Custom hook & better type infer & fix the reliable score
